Some refactors
