+++ Softeq.XToolkit.WhiteLabel.Droid/ImagePicker/DroidImagePickerResult.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.IO;
+﻿using System.IO;
 using System.Threading.Tasks;
 using Android.Graphics;
 using Softeq.XToolkit.WhiteLabel.ImagePicker;
@@ -12,10 +11,11 @@ namespace Softeq.XToolkit.WhiteLabel.Droid.ImagePicker
         public override Task<Stream> GetStream()
         {
             Stream memoryStream = new MemoryStream();
-            if(ImageObject is Bitmap bitmap)
+            if (ImageObject is Bitmap bitmap)
             {
-                var compressFormat = ImageExtension == ImageExtension.Jpg ?
-                    CompressFormat.Jpeg : CompressFormat.Png;
+                var compressFormat = ImageExtension == ImageExtension.Jpg
+                    ? CompressFormat.Jpeg
+                    : CompressFormat.Png;
 
                 bitmap.Compress(compressFormat, (int) (Quality * 100), memoryStream);
             }
