+++ Softeq.XToolkit.WhiteLabel.Droid/ImagePicker/DroidImagePickerService.cs
@@ -12,12 +12,12 @@ namespace Softeq.XToolkit.WhiteLabel.Droid.ImagePicker
     {
         private readonly IPermissionsManager _permissionsManager;
 
+        private TaskCompletionSource<Bitmap> _taskCompletionSource;
+
         public DroidImagePickerService(IPermissionsManager permissionsManager)
         {
             _permissionsManager = permissionsManager;
         }
-
-        private TaskCompletionSource<Bitmap> _taskCompletionSource;
 
         public async Task<ImagePickerResult> PickPhotoAsync(float quality)
         {
