+++ Softeq.XToolkit.WhiteLabel.Droid/ImagePicker/ImagePickerActivity.cs
@@ -1,4 +1,6 @@
-﻿
+﻿// Developed by Softeq Development Corporation
+// http://www.softeq.com
+
 using System;
 using System.Globalization;
 using System.IO;
@@ -14,6 +16,7 @@ using Android.Support.V4.Content;
 using Plugin.CurrentActivity;
 using ImageOrientation = Android.Media.Orientation;
 using Debug = System.Diagnostics.Debug;
+using AUri = Android.Net.Uri;
 
 namespace Softeq.XToolkit.WhiteLabel.Droid.ImagePicker
 {
@@ -26,7 +29,7 @@ namespace Softeq.XToolkit.WhiteLabel.Droid.ImagePicker
 
         private const string ImagesFolder = "Pictures";
 
-        private Android.Net.Uri _fileUri;
+        private AUri _fileUri;
         private Intent _pickIntent;
 
         public static event EventHandler<Bitmap> ImagePicked;
@@ -58,8 +61,9 @@ namespace Softeq.XToolkit.WhiteLabel.Droid.ImagePicker
 
         protected override void OnActivityResult(int requestCode, [GeneratedEnum] Result resultCode, Intent data)
         {
-            Android.Net.Uri uri = null;
+            AUri uri = null;
             Bitmap bitmap = null;
+
             switch (requestCode)
             {
                 case CameraMode:
@@ -69,6 +73,7 @@ namespace Softeq.XToolkit.WhiteLabel.Droid.ImagePicker
                     uri = data?.Data;
                     break;
             }
+
             if (uri != null)
             {
                 bitmap = GetBitmap(uri);
@@ -120,20 +125,17 @@ namespace Softeq.XToolkit.WhiteLabel.Droid.ImagePicker
                     return originalImage;
                 }
 
-                var matrix = new Matrix();
-                matrix.PostRotate(rotation);
-                var rotatedImage = Bitmap.CreateBitmap(originalImage, 0, 0, originalImage.Width, originalImage.Height, matrix, true);
-                originalImage.Recycle();
-                originalImage.Dispose();
-                GC.Collect();
-                return rotatedImage;
+                return RotateImage(originalImage, rotation);
             });
         }
 
         private int GetRotation(ExifInterface exif)
         {
             if (exif == null)
+            {
                 return 0;
+            }
+
             try
             {
                 var orientation = (ImageOrientation) exif.GetAttributeInt(ExifInterface.TagOrientation, (int) ImageOrientation.Normal);
@@ -160,36 +162,47 @@ namespace Softeq.XToolkit.WhiteLabel.Droid.ImagePicker
             }
         }
 
-        private Bitmap GetBitmap(Android.Net.Uri uri)
+        private Bitmap RotateImage(Bitmap originalImage, float degrees)
+        {
+            var matrix = new Matrix();
+            matrix.PostRotate(degrees);
+            var rotatedImage = Bitmap.CreateBitmap(originalImage, 0, 0, originalImage.Width, originalImage.Height, matrix, true);
+            originalImage.Recycle();
+            originalImage.Dispose();
+            GC.Collect();
+            return rotatedImage;
+        }
+
+        private Bitmap GetBitmap(AUri uri)
         {
             Bitmap result = null;
             try
             {
                 result = MediaStore.Images.Media.GetBitmap(CrossCurrentActivity.Current.AppContext.ContentResolver, uri);
             }
-            catch (Java.IO.FileNotFoundException fnfEx)
+            catch (Java.IO.FileNotFoundException ex)
             {
-                Debug.WriteLine("Unable to get Bitmap file from disk " + fnfEx);
+                Debug.WriteLine("Unable to get Bitmap file from disk " + ex);
             }
             return result;
         }
 
-        private System.IO.Stream GetContentStream(Context context, Android.Net.Uri uri)
+        private System.IO.Stream GetContentStream(Context context, AUri uri)
         {
             var stream = System.IO.Stream.Null;
             try
             {
                 stream = context.ContentResolver.OpenInputStream(uri);
             }
-            catch (Java.IO.FileNotFoundException fnfEx)
+            catch (Java.IO.FileNotFoundException ex)
             {
-                Debug.WriteLine("Unable to save picked file from disk " + fnfEx);
+                Debug.WriteLine("Unable to save picked file from disk " + ex);
             }
 
             return stream;
         }
 
-        private Android.Net.Uri GetOutputMediaFile(Context context, string subdir, string name)
+        private AUri GetOutputMediaFile(Context context, string subdir, string name)
         {
             subdir = subdir ?? string.Empty;
 
@@ -215,7 +228,8 @@ namespace Softeq.XToolkit.WhiteLabel.Droid.ImagePicker
                     }
                 }
 
-                return FileProvider.GetUriForFile(context, $"{context.PackageName}.fileprovider", new Java.IO.File(GetUniquePath(mediaStorageDir.Path, name)));
+                return FileProvider.GetUriForFile(context, $"{context.PackageName}.fileprovider",
+                    new Java.IO.File(GetUniquePath(mediaStorageDir.Path, name)));
             }
         }
 
@@ -238,6 +252,5 @@ namespace Softeq.XToolkit.WhiteLabel.Droid.ImagePicker
 
             return System.IO.Path.Combine(folder, nname);
         }
-
     }
 }
