+++ Softeq.XToolkit.WhiteLabel.iOS/ImagePicker/ImagePicker.cs
@@ -6,7 +6,6 @@ using System.Threading.Tasks;
 using CoreGraphics;
 using FFImageLoading;
 using Plugin.Permissions;
-using Plugin.Permissions.Abstractions;
 using Softeq.XToolkit.Common;
 using Softeq.XToolkit.Common.Command;
 using Softeq.XToolkit.Common.Interfaces;
@@ -67,7 +66,7 @@ namespace Softeq.XToolkit.WhiteLabel.iOS.ImagePicker
             Unsubscribe();
 
             var imageKey = _options.AllowEditing ? UIImagePickerController.EditedImage : UIImagePickerController.OriginalImage;
-            var image = ToUpImageOrientation((UIImage)e.Info[imageKey]);
+            var image = ToUpImageOrientation((UIImage) e.Info[imageKey]);
 
             var name = $"{Guid.NewGuid()}.png";
 
@@ -99,7 +98,7 @@ namespace Softeq.XToolkit.WhiteLabel.iOS.ImagePicker
         private async Task<bool> IsPermissionGranted()
         {
             Permissions.PermissionStatus status;
-            if(_options.ImagePickerOpenType == ImagePickerOpenTypes.Camera)
+            if (_options.ImagePickerOpenType == ImagePickerOpenTypes.Camera)
             {
                 status = await _permissionManager.CheckWithRequestAsync<CameraPermission>().ConfigureAwait(false);
             }
@@ -162,7 +161,7 @@ namespace Softeq.XToolkit.WhiteLabel.iOS.ImagePicker
                 view = new UIView(rect);
             });
 
-            var radians = degree * (float)Math.PI / 180;
+            var radians = degree * (float) Math.PI / 180;
             var transformRotate = CGAffineTransform.MakeRotation(radians);
 
             var size = new CGSize();
