+++ Softeq.XToolkit.WhiteLabel.iOS/ImagePicker/IosImagePickerResult.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.IO;
+﻿using System.IO;
 using System.Threading.Tasks;
 using Softeq.XToolkit.WhiteLabel.ImagePicker;
 using UIKit;
@@ -12,13 +11,12 @@ namespace Softeq.XToolkit.WhiteLabel.iOS.ImagePicker
         {
             if (ImageObject is UIImage image)
             {
-                if(ImageExtension == ImageExtension.Jpg)
+                switch (ImageExtension)
                 {
-                    return Task.FromResult(image.AsJPEG(Quality).AsStream());
-                }
-                else if (ImageExtension == ImageExtension.Png)
-                {
-                    return Task.FromResult(image.AsPNG().AsStream());
+                    case ImageExtension.Jpg:
+                        return Task.FromResult(image.AsJPEG(Quality).AsStream());
+                    case ImageExtension.Png:
+                        return Task.FromResult(image.AsPNG().AsStream());
                 }
             }
             return Task.FromResult(Stream.Null);
